I went back to before teaching and cleared up the non-compiling code.

+++ src/main/java/bb_methods_classes_intro/bb_StaticMethod/MyMain.java
@@ -8,7 +8,7 @@ class MyMain{
 //		double cc1 = Math.max( aa, bb );
 		final double cc1 = MyMath.myMax( aa, bb );
 
-		final int cc2 = MyMath.myMax( 0, 2 );
+		final double cc2 = MyMath.myMax( 0, 2 );
 
 		System.out.println( " cc: " + cc1 ) ;
